+++ kubetools_client/kubernetes/job.py
@@ -1,0 +1,99 @@
+import shlex
+from uuid import uuid4
+
+from six import string_types
+
+from .container import make_container_config
+from .util import copy_and_update
+
+
+def make_job_config(
+    config,
+    app_name=None,
+    labels=None,
+    annotations=None,
+    envars=None,
+):
+    '''
+    Builds a Kubernetes job configuration dict.
+    '''
+
+    # We want a copy of these because we'll modify them below
+    labels = labels or {}
+    envars = envars or {}
+    annotations = annotations or {}
+
+    # Generate name
+    job_id = str(uuid4())
+
+    # Attach the ID to labels
+    labels = copy_and_update(labels, {
+        'job-id': job_id,
+    })
+
+    # Figure out the command
+    command = config['command']
+
+    if isinstance(command, string_types):
+        command = shlex.split(command)
+
+    # Get/create description
+    description = config.get('description', 'Run: {0}'.format(command))
+
+    # Attach description to annotations
+    annotations = copy_and_update(annotations, {
+        'description': description,
+    })
+
+    # Update global envars with job specific ones
+    if 'envars' in config:
+        copy_and_update(envars, config['envars'])
+
+    # Make our container
+    container = make_container_config(
+        job_id,
+        {
+            'name': 'upgrade',
+            'command': command,
+            'image': config['image'],
+            'chdir': config.get('chdir', '/'),
+        },
+        envars=envars,
+        labels=labels,
+        annotations=annotations,
+    )
+
+    # Completions default to 1, same as Kubernetes
+    completions = config.get('completions', 1)
+    # Parallelism defaults to completions, also as Kubernetes
+    parallelism = config.get('parallelism', completions)
+
+    return {
+        # Special fields the Kubebuilder pops before pushing - used for updating build
+        # status with the job info.
+        '_description': description,
+        '_app_name': app_name,
+
+        # Normal Kubernetes job config
+        'apiVersion': 'batch/v1',
+        'kind': 'Job',
+        'metadata': {
+            'name': job_id,
+            'labels': labels,
+            'annotations': annotations,
+        },
+        'spec': {
+            'completions': completions,
+            'parallelism': parallelism,
+            'selector': labels,
+            'template': {
+                'metadata': {
+                    'labels': labels,
+                },
+                'spec': {
+                    'restartPolicy': 'Never',
+                    'containers': [container],
+                },
+            },
+        },
+    }